requirements.txt
fastapi[all]
llama-index
llama-index-agent-openai
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-agent-openai-legacy
llama-index-readers-file 
llama-index-vector-stores-chroma
chromadb
llama-index-storage-docstore-mongodb
llama-index-storage-index-store-mongodb
pymongo
motor
python-dotenv